requirements.txt
aiogram==2.25.1
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django~=4.2.2
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
plotly==5.13.0
pyTelegramBotAPI==4.10.0
pytz==2022.7.1
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
tenacity==8.2.1
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2

djangorestframework~=3.14.0
django-filter~=23.2